Add workflow to build/release plugins for OpenSearch
With this workflow for GitHub actions, the plugin will
be built and released when a new tag with the prefix
'osd-' is created. This implies that all the tags
that start with that prefix will build versions for
OpenSearch.

This commit also updates the version of OpenSearch
to the latest. It's neccesary to build the plugin with
the exact dependencies for that version.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>